requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-MCP3008==1.0.2
Adafruit-PureIO==1.1.11
RPi.GPIO==0.7.1
spidev==3.6
